backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.11.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
python-dotenv>=1.0.0
asyncpg>=0.28.0
alembic>=1.12.0
fastapi-cache2>=0.2.1
websockets>=11.0.0
httpx>=0.24.0
yfinance>=0.2.36
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1